fix(analytics): 0.9.1 — temp_c DECIMAL(4,1), so a fractional flag level works
sprint6 SS4/SS5 specify a user-settable temperature flag of 37.5 C over a
37.2-38.1 range. The column was INTEGER, and measured against it every value
in that range rounds to 37 or 38: thresholds 37.2 and 37.9 select identical
rows. The control had two states, not a range, and dragging it across most of
its span did nothing.

INTEGER was defensible on its own terms — the current ring emits whole degrees
and the mapper stores them verbatim, expecting daily averaging to smooth the
discretisation, which it does; baselines were never wrong. It stopped being
defensible when a fractional threshold was specified against it. It also
forecloses the 0.1 C next-gen rings that sprint6 SS7 plans a precision gate
for: the gate would govern the display of data the schema had already
destroyed.

Widening only, so migration 7's copy is lossless and safe to re-run.

Both DDL variants change, not just the local one. The pusher writes local rows
into the remote Iceberg table, so leaving remote at INTEGER would truncate
decimals on push. Cloud sync is off today (local mode), which is the window to
land this; an Iceberg catalog already created from the old DDL still needs a
server-side column evolution before sync is enabled, and the client will not
do it — same rule as migrations 5 and 6.

Two tests carry the reason. One asserts each step of the spec's range selects
a different set; the other that a tenth round-trips. Both fail if the column
returns to INTEGER, verified by doing exactly that.

Separately noted, not changed here: the analytics-core spec declares these
metric columns SMALLINT while the code has always used INTEGER — bpm, hrv_ms,
spo2, stress and temp_c all drift. Benign (wider, no data loss) but it is five
columns of spec disagreeing with code, and it is why the spec's
"temp_c SMALLINT -- whole degrees per firmware" did not read as a constraint
anyone had to reconcile with sprint6.

+++ packages/analytics/src/core/schemas.ts
@@ -108,7 +108,16 @@ export const SCHEMAS: Readonly<Record<TableName, string>> = Object.freeze({
   family_id VARCHAR NOT NULL,
   user_id VARCHAR NOT NULL,
   device_id VARCHAR NOT NULL,
-  temp_c INTEGER NOT NULL
+  -- DECIMAL, not an integer type. The current ring emits whole degrees and
+  -- the mapper stores them verbatim, which made INTEGER look free — but
+  -- sprint6 specifies a user-settable flag level of 37.5 C over a 37.2-38.1
+  -- range, and against an integer column every value in that range rounds to
+  -- 37 or 38. Measured: thresholds 37.2 and 37.9 select identical rows, so
+  -- the control has two states rather than a range. It also forecloses the
+  -- 0.1 C next-gen rings that sprint6 SS7 plans a precision gate for.
+  --
+  -- 4 digits, 1 decimal: -999.9 to 999.9, far past any body temperature.
+  temp_c DECIMAL(4,1) NOT NULL
 ) PARTITIONED BY (day(ts), user_id);`,
 
   activity: `CREATE TABLE activity (
